# [IMP] stock: Speed-up qty_to_order_computed using multi instead of singleton