@robodoo 2025-05-16
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/210437.png)](https://mergebot.odoo.com/odoo/odoo/pull/210437)