@lase-odoo 2025-05-19
@ticodoo, @svs-odoo,  [similar PR](https://github.com/odoo/odoo/pull/201621) assigned to @clesgow.
The reason we group by `location_id` in the current optimization is because certain methods were not written to be called in batch (or at least ar sensitive with respect to the `location_id` of the other unrelated order-points if called in batch) see:
https://github.com/odoo/odoo/blob/6c5b8ef6b61a8f1c4b534a695672ab08c88fe5f8/addons/purchase_stock/models/product.py#L42-L52
https://github.com/odoo/odoo/blob/822e2eb5ac452a2eca48a432253de842815866a2/addons/purchase_stock/models/product.py#L69-L76
More details on https://github.com/odoo/odoo/commit/a58b99509a45a4372aca460eba22893f3358139e where I reverted a batching IMP because of this.